change type used to match recommended policy

+++ test/include/RAJA_test-forall-execpol.hpp
@@ -109,7 +109,7 @@ using CudaForallExecPols = camp::list< RAJA::cuda_exec<128>,
                                        RAJA::cuda_exec_occ_calc<256>,
                                        RAJA::cuda_exec_grid<256, 64>,
                                        RAJA::cuda_exec_explicit<256,2>,
-                                       RAJA::cuda_exec_occ_calc_fraction<256, RAJA::Fraction<int,1,2>> >;
+                                       RAJA::cuda_exec_occ_calc_fraction<256, RAJA::Fraction<size_t,1,2>> >;
 
 using CudaForallReduceExecPols = CudaForallExecPols;
 
@@ -121,7 +121,7 @@ using CudaForallAtomicExecPols = CudaForallExecPols;
 using HipForallExecPols = camp::list< RAJA::hip_exec<128>,
                                       RAJA::hip_exec_occ_calc<256>,
                                       RAJA::hip_exec_grid<256, 64>,
-                                      RAJA::hip_exec_occ_calc_fraction<256, RAJA::Fraction<int,1,2>> >;
+                                      RAJA::hip_exec_occ_calc_fraction<256, RAJA::Fraction<size_t,1,2>> >;
 
 using HipForallReduceExecPols = HipForallExecPols;
